enh: fixed last bugs
Now the algorithms produces the correct values all-around.

The only thing that needs checking is the entropy calculation
of Cold.

Also the variance values are not well-defined. They are not
used either. But they can be used to get bounds on energy intervals
provided the pdf is used to query the smearing of states.

Note that running main checks that the integration of the pdf
is 1 as well as producing comparisons with the old fermid.

This shows that the entropy calculation for the cold distribution
was probably not implemented correctly. This should fix it.